// src/app/models/shop.ts
export interface Shop {
    shopId: string;
        shopname: string;
        shopemail: string;
        shopphone: string;
        contactname: string;
        shopaddr: string;
        shoplocality: string;
        shopType: string;
        lastupdton: Date;
        timestamp;
        docAuthor: string;
        teamId: string;
        shopStatus: boolean; // active shop = true else false;
}